Migration Step 6
Deactivate IDEA plugin "Jangaroo 0.9" (Settings->Plugins).
Install plugins ("Browse Repositories") "Jangaroo 4" and
"Jangaroo Migration to Ext AS 6". As requested, restart IDEA.

Use new menu entry Refactor -> Migrate to Ext AS 6. Enter the
same Ext AS version as in the last step (as of this writing: 6.2.0-6).
This refactors all Ext AS 3.4 API usages to Ext AS 6.2 (in this simple
case: none) and, which is important for this example, replaces
config classes (lower case first letter) by their target classes
(upper case).