Switch to CoreMedia 9 Blueprint Workspace
Now the code has been migrated, but is still in the wrong context:
CM8, not CMS 9. Thus, there are several reference errors (red code).
So this is the right state to move all example files to your new
CMS 9 Blueprint Workspace.

The files remain as-as, so this commit is empty (just documentation).