Extract Object Configuration into MXML
MXML allows to use any ActionScript class that can be instantiated
without arguments. If the class is not part of a component library,
it can be accessed through an ad-hoc namespace derived from its
package name (here: xmlns:js="js.*").
Any writable properties can be assigned via MXML.
Here, we extract the SpeechSynthesisUtterance object from the
ActionScript code. In MXML, helper objects can be created in the
<fx:Declarations> section. The "id" attribute specifies the name to
be used to refer to the object.